pyproject.toml
[project]
name = "pythonproject"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "allure-pytest==2.15.0",
    "allure-python-commons==2.15.0",
    "altgraph==0.17.4",
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "attrs==25.3.0",
    "automat==25.4.16",
    "black==25.9.0",
    "bottle==0.13.4",
    "bottle-websocket==0.2.9",
    "certifi==2025.8.3",
    "cffi==2.0.0",
    "charset-normalizer==3.4.3",
    "click==8.3.0",
    "colorama==0.4.6",
    "constantly==23.10.4",
    "coverage==7.10.6",
    "decorator==5.2.1",
    "et-xmlfile==2.0.0",
    "execnet==2.1.1",
    "faker==37.6.0",
    "fastapi==0.116.1",
    "gevent==25.8.2",
    "gevent-websocket==0.10.1",
    "greenlet==3.2.4",
    "h11==0.16.0",
    "helper==2.5.0",
    "hyperlink==21.0.0",
    "idna==3.10",
    "incremental==24.7.2",
    "iniconfig==2.1.0",
    "jinja2==3.1.6",
    "markupsafe==3.0.2",
    "mypy-extensions==1.1.0",
    "openpyxl==3.2.0b1",
    "outcome==1.3.0.post0",
    "packaging==25.0",
    "pathspec==0.12.1",
    "pefile==2023.2.7",
    "pillow==11.3.0",
    "pip-review==1.3.0",
    "platformdirs==4.4.0",
    "playwright==1.55.0",
    "playwright-stealth==2.0.0",
    "pluggy==1.6.0",
    "pycparser==2.23",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pyee==13.0.0",
    "pygments==2.19.2",
    "pyinstaller==6.16.0",
    "pyinstaller-hooks-contrib==2025.8",
    "pyparsing==3.2.3",
    "pysocks==1.7.1",
    "pytest==8.4.2",
    "pytest-base-url==2.1.0",
    "pytest-black==0.6.0",
    "pytest-cov==7.0.0",
    "pytest-dependency==0.6.0",
    "pytest-html==4.1.1",
    "pytest-metadata==3.1.1",
    "pytest-order==1.3.0",
    "pytest-playwright==0.7.1",
    "pytest-selenium==4.1.0",
    "pytest-variables==3.1.0",
    "pytest-xdist==3.8.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-slugify==8.0.4",
    "pywin32-ctypes==0.2.3",
    "pyyaml==6.0.2",
    "readme-md-docstrings==0.1.0",
    "requests==2.32.5",
    "selenium==4.35.0",
    "setuptools==80.9.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "starlette==0.47.3",
    "tenacity==9.1.2",
    "text-unidecode==1.3",
    "toml==0.10.2",
    "trio==0.30.0",
    "trio-websocket==0.12.2",
    "twisted==25.5.0",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "webdriver-manager==4.0.2",
    "websocket-client==1.8.0",
    "wsproto==1.2.0",
    "zope-event==5.1.1",
    "zope-interface==7.2",
]
